Adaptation Account List

+++ app/src/main/java/me/ash/reader/data/repository/AccountRepository.kt
@@ -2,6 +2,7 @@ package me.ash.reader.data.repository
 
 import android.content.Context
 import dagger.hilt.android.qualifiers.ApplicationContext
+import kotlinx.coroutines.flow.Flow
 import me.ash.reader.R
 import me.ash.reader.data.dao.AccountDao
 import me.ash.reader.data.dao.ArticleDao
@@ -12,6 +13,8 @@ import me.ash.reader.data.model.account.AccountType
 import me.ash.reader.data.model.group.Group
 import me.ash.reader.ui.ext.currentAccountId
 import me.ash.reader.ui.ext.getDefaultGroupId
+import me.ash.reader.ui.ext.showToast
+import me.ash.reader.ui.ext.showToastLong
 import javax.inject.Inject
 
 class AccountRepository @Inject constructor(
@@ -22,6 +25,9 @@ class AccountRepository @Inject constructor(
     private val feedDao: FeedDao,
     private val articleDao: ArticleDao,
 ) {
+    fun getAccounts(): Flow<List<Account>> = accountDao.queryAllAsFlow()
+
+    fun getAccountById(accountId: Int): Flow<Account?> = accountDao.queryAccount(accountId)
 
     suspend fun getCurrentAccount(): Account = accountDao.queryById(context.currentAccountId)!!
 
@@ -55,11 +61,16 @@ class AccountRepository @Inject constructor(
     }
 
     suspend fun delete(accountId: Int) {
+        if (accountDao.queryAll().size == 1) {
+            context.showToast(context.getString(R.string.must_have_an_account))
+            return
+        }
         accountDao.queryById(accountId)?.let {
             articleDao.deleteByAccountId(accountId)
             feedDao.deleteByAccountId(accountId)
             groupDao.deleteByAccountId(accountId)
             accountDao.delete(it)
+            context.showToastLong(context.getString(R.string.delete_account_toast))
         }
     }
 }
